+++ core-mit/src/main/java/eth/likespro/commons/models/Entity.kt
@@ -1,0 +1,5 @@
+package eth.likespro.commons.models
+
+interface Entity<ID: Any> {
+    val id: ID
+}